PowerHAL: init with performance profile
When PowerHAL is initialised, it defaults to the balanced profile.
That can slow down the boot especially when an optimzation is running.
To avoid that, initialise PowerHAL with the performance profile.
To avoid getting stuch in that profile, add a thread that detects if bootanimation is running.
The thread will periodically check if bootanimation is running, and will set the normal profile once done.

Change-Id: I57fd9693900d68190304d486dce2d17b375815ed